refactor: split omp 3SPN2 Base-Base into BP and CS
