+++ mjolnir/omp/omp.hpp
@@ -29,7 +29,8 @@
 #include <mjolnir/omp/GlobalPairInteraction.hpp>
 #include <mjolnir/omp/GlobalPairExcludedVolumeInteraction.hpp>
 #include <mjolnir/omp/GlobalPairLennardJonesInteraction.hpp>
-#include <mjolnir/omp/ThreeSPN2BaseBaseInteraction.hpp>
+#include <mjolnir/omp/ThreeSPN2BasePairInteraction.hpp>
+#include <mjolnir/omp/ThreeSPN2CrossStackingInteraction.hpp>
 #include <mjolnir/omp/PWMcosInteraction.hpp>
 #include <mjolnir/omp/PositionRestraintInteraction.hpp>
 #include <mjolnir/omp/ExternalDistanceInteraction.hpp>
